all: add grouping; change filtering behaviour
This is a big commit containing a big refactoring due to introduction
of groups.

Previous to groups the raw measurement labels were used everywhere
in the application to refer to individual measurements. Groups now
combine several single measurements together to a single identity.
In the database layer we now have to translate incoming groups to
individual measurements and this layer only needs to now know the
relationship between groups and measuerments.  The access control
ayer was simplified and moved directly to the database layer. Since
data access control is done for public users and for measurements,
there is no need to have an extra layer for that. The change to
groups had also an effect on the user interface layer. Moreover the
commit contains several smaller fixes and changes.

This all should have been several commits.

Related https://gitlab.inf.unibz.it/lter/browser/-/issues/98
Related https://gitlab.inf.unibz.it/lter/browser/-/issues/91